refactor database out
