+++ rope/contrib/autoimport/sqlite.py
@@ -1,33 +1,27 @@
 """AutoImport module for rope."""
-
-import contextlib
-import json
-from hashlib import sha256
-import secrets
-import re
 import sqlite3
+import contextlib
+import re
 import sys
 import warnings
 from collections import OrderedDict
 from concurrent.futures import Future, ProcessPoolExecutor, as_completed
-from datetime import datetime
 from itertools import chain
 from pathlib import Path
-from threading import local
 from typing import Generator, Iterable, Iterator, List, Optional, Set, Tuple
-
-from rope.base import exceptions, libutils, resourceobserver, taskhandle, versioning
+from rope.contrib.autoimport._database import Database
+from rope.base import exceptions, libutils, resourceobserver, taskhandle
 from rope.base.project import Project
 from rope.base.resources import Resource
 from rope.contrib.autoimport import models
 from rope.contrib.autoimport.defs import (
     ModuleFile,
-    Name,
     NameType,
     Package,
     PackageType,
     SearchResult,
     Source,
+    Name,
 )
 from rope.contrib.autoimport.parse import get_names
 from rope.contrib.autoimport.utils import (
@@ -80,7 +74,7 @@ class AutoImport:
 
     """
 
-    connection: sqlite3.Connection
+    database: Database
     memory: bool
     project: Project
     project_package: Package
@@ -128,81 +122,12 @@ class AutoImport:
                 "`AutoImport(memory=True)` explicitly.",
                 DeprecationWarning,
             )
-        self.thread_local = local()
-        self.connection = self.create_database_connection(
-            project=project,
-            memory=memory,
-        )
-        self._setup_db()
+        self.database = Database(self.project, self.memory)
         if observe:
             observer = resourceobserver.ResourceObserver(
                 changed=self._changed, moved=self._moved, removed=self._removed
             )
             project.add_observer(observer)
-
-    @classmethod
-    def create_database_connection(
-        cls,
-        *,
-        project: Optional[Project] = None,
-        memory: bool = False,
-    ) -> sqlite3.Connection:
-        """
-        Create an sqlite3 connection
-
-        project : rope.base.project.Project
-            the project to use for project imports
-        memory : bool
-            if true, don't persist to disk
-        """
-
-        def calculate_project_hash(data: str) -> str:
-            return sha256(data.encode()).hexdigest()
-
-        if not memory and project is None:
-            raise Exception("if memory=False, project must be provided")
-        if memory or project is None or project.ropefolder is None:
-            # Allows the in-memory db to be shared across threads
-            # See https://www.sqlite.org/inmemorydb.html
-            project_hash: str
-            if project is None:
-                project_hash = secrets.token_hex()
-            elif project.ropefolder is None:
-                project_hash = calculate_project_hash(project.address)
-            else:
-                project_hash = calculate_project_hash(project.ropefolder.real_path)
-            return sqlite3.connect(
-                f"file:rope-{project_hash}:?mode=memory&cache=shared", uri=True
-            )
-        else:
-            return sqlite3.connect(project.ropefolder.pathlib / "autoimport.db")
-
-    @property
-    def connection(self):
-        """
-        Creates a new connection if called from a new thread.
-
-        This makes sure AutoImport can be shared across threads.
-        """
-        if not hasattr(self.thread_local, "connection"):
-            self.thread_local.connection = self.create_database_connection(
-                project=self.project,
-                memory=self.memory,
-            )
-        return self.thread_local.connection
-
-    @connection.setter
-    def connection(self, value: sqlite3.Connection):
-        self.thread_local.connection = value
-
-    def _setup_db(self):
-        models.Metadata.create_table(self.connection)
-        version_hash = list(
-            self._execute(models.Metadata.objects.select("version_hash"))
-        )
-        current_version_hash = versioning.calculate_version_hash(self.project)
-        if not version_hash or version_hash[0][0] != current_version_hash:
-            self.clear_cache()
 
     def import_assist(self, starting: str):
         """
@@ -218,7 +143,7 @@ class AutoImport:
         __________
         Return a list of ``(name, module)`` tuples
         """
-        results = self._execute(
+        results = self.database._execute(
             models.Name.import_assist.select("name", "module", "source"), (starting,)
         ).fetchall()
         return sort_and_deduplicate_tuple(
@@ -282,7 +207,7 @@ class AutoImport:
         """
         if not exact_match:
             name = name + "%"  # Makes the query a starts_with query
-        for import_name, module, source, name_type in self._execute(
+        for import_name, module, source, name_type in self.database._execute(
             models.Name.search_by_name_like.select("name", "module", "source", "type"),
             (name,),
         ):
@@ -305,7 +230,7 @@ class AutoImport:
         """
         if not exact_match:
             name = name + "%"  # Makes the query a starts_with query
-        for module, source in self._execute(
+        for module, source in self.database._execute(
             models.Name.search_submodule_like.select("module", "source"), (name,)
         ):
             parts = module.split(".")
@@ -322,7 +247,7 @@ class AutoImport:
                     NameType.Module.value,
                 )
             )
-        for module, source in self._execute(
+        for module, source in self.database._execute(
             models.Name.search_module_like.select("module", "source"), (name,)
         ):
             if "." in module:
@@ -333,20 +258,14 @@ class AutoImport:
 
     def get_modules(self, name) -> List[str]:
         """Get the list of modules that have global `name`."""
-        results = self._execute(
+        results = self.database._execute(
             models.Name.search_by_name.select("module", "source"), (name,)
         ).fetchall()
         return sort_and_deduplicate(results)
 
     def get_all_names(self) -> List[str]:
         """Get the list of all cached global names."""
-        return self._execute(models.Name.objects.select("name")).fetchall()
-
-    def _dump_all(self) -> Tuple[List[Name], List[Package]]:
-        """Dump the entire database."""
-        name_results = self._execute(models.Name.objects.select_star()).fetchall()
-        package_results = self._execute(models.Package.objects.select_star()).fetchall()
-        return name_results, package_results
+        return self.database._execute(models.Name.objects.select("name")).fetchall()
 
     def generate_cache(
         self,
@@ -365,9 +284,7 @@ class AutoImport:
         job_set = task_handle.create_jobset(
             "Generating autoimport cache", len(resources)
         )
-        self._execute(
-            models.Package.delete_by_package_name, (self.project_package.name,)
-        )
+        self.database.delete_package(self.project_package.name)
         futures = []
         with ProcessPoolExecutor() as executor:
             for file in resources:
@@ -375,9 +292,9 @@ class AutoImport:
                 module = self._resource_to_module(file, underlined)
                 futures.append(executor.submit(get_names, module, self.project_package))
         for future in as_completed(futures):
-            self._add_names(future.result())
+            self.database.add_names(future.result())
             job_set.finished_job()
-        self.connection.commit()
+        self.database.commit()
 
     def generate_modules_cache(
         self,
@@ -406,23 +323,23 @@ class AutoImport:
         packages = list(filter_packages(packages, underlined, existing))
         if not packages:
             return
-        self._add_packages(packages)
+        self.database.add_packages(packages)
         job_set = task_handle.create_jobset("Generating autoimport cache", 0)
         if single_thread:
             for package in packages:
                 for module in get_files(package, underlined):
                     job_set.started_job(module.modname)
                     for name in get_names(module, package):
-                        self._add_name(name)
+                        self.database.add_name(name)
                         job_set.finished_job()
         else:
             for future_name in as_completed(
                 get_future_names(packages, underlined, job_set)
             ):
-                self._add_names(future_name.result())
+                self.database.add_names(future_name.result())
                 job_set.finished_job()
 
-        self.connection.commit()
+        self.database.commit()
 
     def _get_packages_from_modules(self, modules: List[str]) -> Iterator[Package]:
         for modname in modules:
@@ -438,13 +355,12 @@ class AutoImport:
 
     def close(self):
         """Close the autoimport database."""
-        self.connection.commit()
-        self.connection.close()
+        self.database.close()
 
     def get_name_locations(self, name):
         """Return a list of ``(resource, lineno)`` tuples."""
         result = []
-        modules = self._execute(
+        modules = self.database._execute(
             models.Name.search_by_name_like.select("module"), (name,)
         ).fetchall()
         for module in modules:
@@ -461,34 +377,6 @@ class AutoImport:
                         if resource is not None and lineno is not None:
                             result.append((resource, lineno))
         return result
-
-    def clear_cache(self):
-        """Clear all entries in global-name cache.
-
-        It might be a good idea to use this function before
-        regenerating global names.
-
-        """
-        with self.connection:
-            self._execute(models.Name.objects.drop_table())
-            self._execute(models.Package.objects.drop_table())
-            self._execute(models.Metadata.objects.drop_table())
-            models.Name.create_table(self.connection)
-            models.Package.create_table(self.connection)
-            models.Metadata.create_table(self.connection)
-            data = (
-                versioning.calculate_version_hash(self.project),
-                json.dumps(versioning.get_version_hash_data(self.project)),
-                datetime.utcnow().isoformat(),
-            )
-            assert models.Metadata.columns == [
-                "version_hash",
-                "hash_data",
-                "created_at",
-            ]
-            self._execute(models.Metadata.objects.insert_into(), data)
-
-            self.connection.commit()
 
     def find_insertion_line(self, code):
         """Guess at what line the new import should be inserted."""
@@ -516,9 +404,9 @@ class AutoImport:
         module = self._resource_to_module(resource, underlined)
         self._del_if_exist(module_name=module.modname, commit=False)
         for name in get_names(module, self.project_package):
-            self._add_name(name)
+            self.database.add_name(name)
         if commit:
-            self.connection.commit()
+            self.database.commit()
 
     def _changed(self, resource):
         if not resource.is_folder():
@@ -531,9 +419,9 @@ class AutoImport:
             self.update_resource(newresource)
 
     def _del_if_exist(self, module_name, commit: bool = True):
-        self._execute(models.Name.delete_by_module_name, (module_name,))
+        self.database._execute(models.Name.delete_by_module_name, (module_name,))
         if commit:
-            self.connection.commit()
+            self.database.commit()
 
     def _get_python_folders(self) -> List[Path]:
         def filter_folders(folder: Path) -> bool:
@@ -566,13 +454,11 @@ class AutoImport:
                 packages.append(package_tuple)
         return packages
 
-    def _add_packages(self, packages: List[Package]):
-        data = [(p.name, str(p.path)) for p in packages]
-        self._executemany(models.Package.objects.insert_into(), data)
-
     def _get_packages_from_cache(self) -> List[str]:
         existing: List[str] = list(
-            chain(*self._execute(models.Package.objects.select_star()).fetchall())
+            chain(
+                *self.database._execute(models.Package.objects.select_star()).fetchall()
+            )
         )
         existing.append(self.project_package.name)
         return existing
@@ -583,27 +469,7 @@ class AutoImport:
             self._del_if_exist(modname)
 
     def _add_future_names(self, names: Future):
-        self._add_names(names.result())
-
-    @staticmethod
-    def _convert_name(name: Name) -> tuple:
-        return (
-            name.name,
-            name.modname,
-            name.package,
-            name.source.value,
-            name.name_type.value,
-        )
-
-    def _add_names(self, names: Iterable[Name]):
-        if names is not None:
-            self._executemany(
-                models.Name.objects.insert_into(),
-                [self._convert_name(name) for name in names],
-            )
-
-    def _add_name(self, name: Name):
-        self._execute(models.Name.objects.insert_into(), self._convert_name(name))
+        self.database.add_names(names.result())
 
     def _find_package_path(self, target_name: str) -> Optional[Package]:
         if target_name in sys.builtin_module_names:
@@ -638,10 +504,15 @@ class AutoImport:
             resource_path.name == "__init__.py",
         )
 
-    def _execute(self, query: models.FinalQuery, *args, **kwargs):
-        assert isinstance(query, models.FinalQuery)
-        return self.connection.execute(query._query, *args, **kwargs)
+    def _dump_all(self) -> Tuple[List[Name], List[Package]]:
+        """Dump the entire database."""
+        return self.database._dump_all()
 
-    def _executemany(self, query: models.FinalQuery, *args, **kwargs):
-        assert isinstance(query, models.FinalQuery)
-        return self.connection.executemany(query._query, *args, **kwargs)
+    def clear_cache(self):
+        """Clear all entries in global-name cache.
+
+        It might be a good idea to use this function before
+        regenerating global names.
+
+        """
+        self.database.clear_cache()
